pyproject.toml
[tool.poetry]
name = "wishlist-magalu"
version = "0.1.0"
description = ""
authors = ["Helton Teixeira de Souza <heltonteixeiradesouza@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"
pillow = "^11.0.0"

[tool.poetry.group.dev.dependencies]
python-decouple = "^3.8"
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
ruff = "^0.7.3"
pytest-django = "^4.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 91
extend-exclude = ['migrations', 'manage.py']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = 'core.settings'
python_files = ['test_*.py', '*_test.py']
addopts = '-p no:warnings --cov=.'